fix(math.nt): overflow guards, sqrt clamp, non-residue check
- MoebiusPrefix: cast (long)i*primes[j] in linear-sieve guard (was:
  int overflow -> negative composite index -> OOB write).
- PowerfulNumbers.BuildRawPowerful: clamp maxA down while
  maxA*maxA*b3 > limit (was: Math.Sqrt overshoot emitted a powerful
  number strictly greater than limit).
- PrimeFactorPower.Run(long,...): widen outPrimes to long* (was: int*
  truncated factors p>int.MaxValue and the large prime cofactor n).
  Test updated: int* -> long*, added factor assertions for 60.
- SquareFree.Prefix: long loop bound i*i<=n (was: int i*i overflow ->
  negative i2 -> negative-index write).
- SquareFree.Count: floor(isqrt(n)) via clamp loop + widen limit to
  long before narrowing (was: (int)Math.Sqrt overflows for n near
  long.MaxValue).
- SquareFree.SieveMobius: (long)i*primes[j] guard (same overflow class).
- TonelliShanks.ModSqrtOne p==3 (mod 4) fast path: verify r*r==a and
  return -1 for quadratic non-residues (was: returned bogus root,
  inconsistent with the general path which correctly returns -1).

Reviewed and assessed false positive:
- RandomShuffle.Run: RandomInt.Next is Lemire (hash*bound>>32), always
  non-negative; the shuffle index j in [0,i] is correct.

+++ src/IAFahim.Math.NT/PrimeFactorPower.cs
@@ -21,7 +21,7 @@ namespace IAFahim.Math.NT
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static int Run(long n, int* outPrimes, int* outExponents)
+        public static int Run(long n, long* outPrimes, int* outExponents)
         {
             int count = 0;
             for (long p = 2; p * p <= n; p++)
@@ -29,13 +29,13 @@ namespace IAFahim.Math.NT
                 if (n % p != 0) continue;
                 int e = 0;
                 while (n % p == 0) { n /= p; e++; }
-                outPrimes[count] = (int)p;
+                outPrimes[count] = p;
                 outExponents[count] = e;
                 count++;
             }
             if (n > 1)
             {
-                outPrimes[count] = (int)n;
+                outPrimes[count] = n;
                 outExponents[count] = 1;
                 count++;
             }
